Wrap memory tracker sender in handle struct
Summary: Later diffs (starting with the next diff) add more channels for things to talk to the memory tracker. It's convenient to store them in a single handle to have one things to pass around. This moves the existing Sender into the struct but doesn't add anything else to it.

Reviewed By: Will-MingLun-Li

Differential Revision: D81165563

fbshipit-source-id: 1ed842c00ea953891e74922b5a495e5c8fd4048f