Expose optimized artifacts to archive link style
Summary: Note that we are using `pic` in optimized and debuggable transformations here.  Ideal way to set it up would be to expose whole matrix of debuggable/optimized/pic instead.  But it increases retained memory (not much really, but still) so in the mean time I decied to not do it. As a downside, if transformation spec is used to build a binary statically without -fPIE, we are still using pic objects for it. I haven't found any real world impact of that though. Another point is: transformation specs are not used for building for production and only affect local development so we can do this experiment first. I may revisit that decision.

Differential Revision: D80709936

fbshipit-source-id: b99fb446d7e478c8f7f94e6a6ebd38079190c070