Flatten TSet if transformation provider is passed
Summary:
By default linkable for binaries are collected from `TSet`. The downside is that they are *predetermined*, e.g. there is no way to add any conditions to the traversal using `TSet` projections.
To workaround this we can iterate over `TSet` the same way projection would do it but then we are able to add conditions to decide which flavor to pick for each linkable (of `cxx_library`) we depend on. For now this just an unchanged copy of `projection` implementation. We will actually add conditions based on transformations in further diffs

Differential Revision: D78023445

fbshipit-source-id: 9d655e66a5d150743adfd5142e80dffe9aa75bab