Don't try to verify that RE uses the same persistent worker in tests
Summary:
Per the inline comment, this check causes the test to be flaky.

Now, we are verifying that (1) it ran remotely, (2) it used a persistent worker, but not that the persistent worker is the same for both actions. I think that's the best we can do while guaranteeing that the test is not flaky.

Reviewed By: jiawei-lyu

Differential Revision: D81489430

fbshipit-source-id: 208fdd780d3eb72f8bcf549a20bedec38bdd2b3f