Add env to WorkerInfo
Summary:
We want to be able to use an `env` just for the remote worker initialization. This is important - if we don't have this, then we either can't provide an `env` to the worker initialization, or we have to use the `env` from the action, and that might change across actions, meaning that we don't get a cache hit from the worker itself and so don't benefit from the worker infrastructure correctly.

I tried to get DevMate to do this for me, since this starlark stuff is kinda tricky, but it didn't do a very good job so I just shamelessly copied most of this from https://www.internalfb.com/code/fbsource/fbcode/buck2/app/buck2_build_api/src/interpreter/rule_defs/provider/builtin/external_runner_test_info.rs

Reviewed By: christolliday

Differential Revision: D80804738

fbshipit-source-id: 22d2221b5d09a6fde298f471cce4cf3bdebd092c