apple_library_for_distribution
Summary:
The Problem
==========
Using BUCK to ship dylibs that are intended for bundling inside Frameworks / Xcframeworks is currently problematic. There are a few things that work:
- Shipping a single dylib, exposing a single target alsongside all it's dependencies (which naturally should be tightly controlled).
- Shipping one or more static archives, controlling the embedded deps via [apple_static_archive](https://www.internalfb.com/code/fbsource/[d4db04de673c155a61f9c016d37301601c550a76]/fbcode/buck2/prelude/apple/apple_static_archive.bzl?lines=18)

What doesn't currently work:
- Shipping more than one dylibs, where there are intersecting dependency graphs
  - This is because each required symbol must be defined in one and only one dylib, and there is no way to use link groups to control this without losing the headers and other files that apple_binary needs to package into the Xcframework
- Using apple_library() for other purposes that modify an underlying apple_library() (e.g. inclusion of Xplugins), as this again loses the headers from the original apple_library()

This diff
=======
This diff adds a new rule, `apple_library_for_distribution`. The aim is to behave identically to `apple_library`, and so apply all the link-group logic etc. for packaging up the dylibs. However, it also accepts a `distribution_dep` dependency, and the providers will forward on the `public_framework_headers` and `swiftmodule` contents from that target.

As a result, this target can be used to create complex sets of interdependent Xcframeworks via Buck, or include generated code like Xplugins.

Reviewed By: milend

Differential Revision: D81452188

fbshipit-source-id: 9bc836ace18498cf1528be6d185b58cceafb40ca