Turn on content-based path hashing for android_resource asset dirs
Summary:
This was initially turned on in D76336755, but it was reverted due to some Windows build failures (see T227133266).

Those failures were due to some paths now being too long. By turning on content-based paths for the things that the symlinks are pointing at, we avoid the paths being too long, so that's what this diff does in addition.

Reviewed By: dvir, jselbo

Differential Revision: D80705990

fbshipit-source-id: d2cfd3fd490153fa3ebe0c52d788e55633dba637