Prevent local hrl files from breaking hermeticity
Summary:
# Context
When `erlc` tries to resolve an `-include(...).` it will search in the directory of the file being compiled before trying the directories specified with `-I`

# Problem
Because currently we run `erlc` on the actual `.erl` file to build, it is possible that sibling `.hrl` files will be taken into account, even if they were not explicitly listed in the `srcs` attr of the `erlang_app()`. This breaks hermeticity of local builds.

# This diff
We now put all srcs in a linked "src" directory, and run "erlc" on those files. We then link to that directory when building the app structure, when include_src is true

Reviewed By: michalmuskala

Differential Revision: D81443303

fbshipit-source-id: 98c9c2d2d1a4ddfe1605e6518f894f856a5f0850