If set, use WorkerInfo env when initializing local workers
Summary:
Right now, we just use the env of the first action that runs with the worker. Instead, if the worker sets an env, then use that.

Right now, no worker sets an env so this is a no-op. We can migrate use cases and then just unconditionally use the worker env.

Reviewed By: christolliday

Differential Revision: D80804735

fbshipit-source-id: 649abfa7be1a86fddc125f218d86b0d4e7246284