Fix realease cgroup issue
Summary:
The current impl realse the cgroup after `to_tonic().await`. It is wrong, since it return a stream and after the await, the process is still running.

In this diff, we inspect on the stream when getting the exist event, we realse the cgroup.
Since block need a static lifetime, we expose `PoolState` so that we can have `Arc<Mutext<>>` for it and use it in the block.

Reviewed By: christolliday

Differential Revision: D81736739

fbshipit-source-id: 9293d460be75f0c75c9fcb51f7f02219ff5c4c05