Fix --exit-when=notidle handling
Summary:
This change fixes a slight bug in the original implementation of the `--exit-when=notidle` flag. The intention of the `notidle` option is to prevent an incoming command from executing if the Buck daemon is busy with another command already.

With the original change we were actualy enforcing the flag if the state of the new command was different, *__even if the daemon was idle__*. That's incorrect. As an example:

```
# Build target with a config modifier
$ buck2 build -c foo.bar=true //my:binary

# Build target normally, but with --exit-when=notidle and get a not idle error
# even though the daemon is not busy.
$ buck2 build //my:binary --exit-when=notidle
Buck daemon is busy processing another command:

Caused by:
    `--exit-when=notidle` was set, and buck daemon is not idle.

```

The correct implementation for the flag is to check for active commands early and exit if there are any. Otherwise, proceed as normal.

Reviewed By: JakobDegen

Differential Revision: D81810786

fbshipit-source-id: d2eae18955c23e68ec87a921449a307de97d37a1