Send cgroup path to daemon
Summary:
To track peak memory for an execution action, the daemon needs to associate running actions with cgroup paths. Using a cgroup pool means it can't work out the path ahead of time so it needs to be sent from the forkserver.

This propagates it via ProcessCommand and ProcessGroup since that simplifies only sending it when and if the process using the cgroup has successfully spawned.

Added a temporary debug statement in the place where this value is being passed to and where it will be used later.

Reviewed By: Will-MingLun-Li

Differential Revision: D81393096

fbshipit-source-id: c501f24ed834cedd1a8c7a7be73303a63c63c855