Add ActionCgroups to memory tracker
Summary:
Using a cgroup pool means memory.peak will report the peak for all actions that have previously used that cgroup. To get the peak memory we need to track when each action starts and stops and poll memory.current while they run.

If any errors are encountered while polling a specific action, they will be stored and the last error will be reported as a soft error when the action finishes, instead of continuously reporting errors while it runs.

We need a similar mechanism to associate running actions with cgroups in order to freeze them and report which actions were frozen, so this same component will be used to implement action freezing in a later diff.

Reviewed By: Will-MingLun-Li

Differential Revision: D81466652

fbshipit-source-id: 171d03736c95edfd6c0cc1c5193d81465900748c