Use new thread for memory tracker
Summary:
Currently the memory tracker loop runs on the server `buck2-tn` tokio runtime used for client IO, which has 2 worker threads that we should be very careful to not block.
We should be using a dedicated thread for this already, but seems especially important with later changes to poll all cgroups and freeze them which involves blocking IO to cgroup files.

Reviewed By: Nero5023

Differential Revision: D81632135

fbshipit-source-id: 272deeb27cc6a79d867afb55f0c45ac6797a051e