buck: cxx: add ctx.attr.extra_dwp_flags and allow per-directory rollout to disable llvm-dwp --continue-on-cu-index-overflow (try 2)
Summary:
Unicorn uses `folly::symbolizer` at runtime to print stack traces or log stack traces to scuba.
`folly::symbolizer` may crash on corrupt DWARF/DWP leading to incidents or time wasted investigating pre-release crashes.

I tried to remove `--continue-on-cu-index-overflow` with the plan that broken builds can be onboarded to link groups, but there still are projects that can't use link groups (and don't mind corrupt debug info).
- {D63291887}

Add a per-directory rollout (onboard `unicorn`, `search`, `nlp_tools` which I know use `folly::symbolizer` at runtime) and remove `--continue-on-cu-index-overflow` for those directories and ensure `llvm-dwp` either builds correct DWP files or exits with an error.

 ---

For future reference and aid for others that hit DWP corruption issues, fixes involved turning on link groups:
- {D63277289}
- {D81615968}

 ---
Attempt #1 removed the default `--continue-on-cu-index-overflow` from `dwp_wrapper.py` and then tried to add it back explicitly, but it was done in a manner that missed some cases that end up calling into `dwp`.

This attempt keeps the default in `dwp_wrapper.py` (so should not cause other build issues) and adds an override for the codepaths that matter to `unicorn` & co. but keeps the default `--continue-on-cu-index-overflow`.

#forcetdhashing

Reviewed By: mzlee

Differential Revision: D81839152

fbshipit-source-id: 92dcdd76b2cd0e255eeabf3d76648c39dade1a36