Add BUCK2_HARD_ERROR=panic option for soft errors
Summary:
We should test that soft_errors are not hit during e2e tests, even if they happen on background threads and aren't propagated up to cause the buck invocation to fail, since this is a pretty common place to use soft errors.

Specifically want to add this to check that no soft errors are fired during action freezing.

This causes a bunch of e2e tests to fail, for now making it opt-in and adding a soft error allowlist so that at least resource control tests can use it.

Reviewed By: Will-MingLun-Li

Differential Revision: D81886603

fbshipit-source-id: c414b91e5ad4537eca5db28e026b022a0d3a04bb