Add simple action freezing
Summary:
Initial implementation of action freezing. This behaves similarly to the existing local action throttle in that it will freeze all but a single action once pressure is detected, and unfreeze actions one at a time.

This should be changed later to only freezes as many actions as are needed to relieve memory pressure but this should be enough to test if we can reliably prevent OOMs.

Reviewed By: Nero5023

Differential Revision: D81503298

fbshipit-source-id: 250980248fccb18027a2900b85cbb10f267a78fd