Add incremental_state_table implementation
Summary:
Unused for now, but added code for creating, inserting, reading, and deleting from the sqlite table that will be used to store the incremental states.

Mostly followed materializer table. Some details may change in the following diffs depending on what's needed but this is the bulk of the implementation

Reviewed By: IanChilds

Differential Revision: D80668128

fbshipit-source-id: 4479919cca7ae2274d8f090e2d37d4ee06b42fdd