Use and store Incremental DB State in Sqlite
Summary:
Update the in-memory only incremental state to also use the state read from Sqlite if enabled (It's enabled by default), if `sqlite_incremental_state` is set to false, then the state will still be stored only in-memory.

The state is shared within the daemon and concurrency is being handled by DashMap and Arc. Insertion/Deletion to the state/database are only done at the end of an action execution, I'm logging a soft error for failures right now, will monitor the frequency and error types to determine if we should be logging an actual error for these

Reviewed By: scottcao

Differential Revision: D80991055

fbshipit-source-id: 3f143f70df1e7b578342cf590579e020b7ba45bc