Add support for ignoring tags in incremental metadata
Summary:
The RFC is here: https://fb.workplace.com/groups/buck2dev/posts/4100582480229793/

There is some discussion on it from JakobDegen but this is blocking Kotlin folks trying to roll out incremental KotlinCD so trying to progress this, I think we could change it later if need be anyway.

This diff mostly just wires up what the previous diffs enabled - the `visitor` now easily gives us the ability to filter inputs by artifact tag, so that's what we do.

Reviewed By: cbnilrem

Differential Revision: D81779838

fbshipit-source-id: b5ae82f3ece2136d20e97154fb6a0b32f0ec1d39