Diagnostics: add `[check]` subtarget on `apple_bundle()`
Summary:
Adds a `[check]` subtarget on `apple_bundle()` which is a transitive check on all `apple_library()`, `apple_binary()` and `apple_bundle()` targets in the dep tree.

This enables a workflow where the whole app can be typechecked without having to link and build, speeding up the iteration process when not requiring to run the app or check for linker errors.

Reviewed By: IanChilds

Differential Revision: D82103172

fbshipit-source-id: 1af20913e7f1f79ca84b07018d1e9101a027669a