Introduce helper method getSecondaryDexName
Summary: Currently, `getSecondaryDexJarName` returns a string of the form `%s-%d.dex.jar`. However, I'd like to have the option to return a string of the form `%s-%d.dex` (in case we don't apply JAR compression). This diff renames the method to `getSecondaryDexName` and introduces an argument `suffix`. That way, we can control the suffix. While that doesn't matter in the current version of the code, this is exactly what I need in my next diff. The behavior is unchanged.

Reviewed By: IanChilds

Differential Revision: D82406272

fbshipit-source-id: aa5e3849025183da32dbf8e9cd421b067a41acd2