Rename output_artifact -> output_artifacts
Summary:
We want it to be 16 chars so that it can be easily rewritten if it needs to be.

Meta: The action digest test is going to fail, but this only affects a few action digests, and I also roll out buck2 version bumps, so I'm not going to use the buckconfig approach which would be super-tedious and invasive.

Reviewed By: JakobDegen

Differential Revision: D82299612

fbshipit-source-id: 496edd37497547d12d7eadfdca1e2859cbf0491c