Add support for raw_subdir compression to SecondaryDexCompressionExecutableMain
Summary:
This diff introduces support for `compression = "raw_subdir"` (D82302583) to the secondary dex compression CLI. This Java CLI is used as an alternative for the code written in Starlark.

The logic and output is the same as in D82302583 - except written in Java. See that diff for more details.

Reviewed By: IanChilds

Differential Revision: D82406334

fbshipit-source-id: 53988eeeebce592275f806515d25b8795af3936b