Track max DICE in-progress and compute keys in recorder
Summary:
# This Stack

I want to log more state out of Buck2. The reason I'm doing this is that we are trying a new approach for feature exploration for Citadel sizing, which consists of using "after-the-fact" data (like this) to see if it'd be useful as a predictor of OOMs (and if it is, then we can try to walk to "before-the-fact" data that we could actually use in predictions).

In other words, I'd like to find out if letting the model know what peak concurrency to expect would help it make better predictions (by looking at how high the retries peaked).

If it does, then I'll see if I can walk that back to some features that are available earlier (or it might motivate some work around smoothing the peaks).

This stack was generated by Devmate, but I did review the diffs before hitting publish.

# This Diff

Track count of DICE keys in in-progress and in-compute stages.

Reviewed By: IanChilds

Differential Revision: D82301780

fbshipit-source-id: dcfc577a1546c7360dc679bff219826b3b5b0cdc