artifacts: Use `ValueTyped` for `outputs_for_error_handler`
Summary:
So, first of all, cloning the `StarlarkOutputArtifact`s here isn't really great - this probably should have been using `ValueTypedComplex` like nearby code

This diff instead though updates it to `ValueTyped`, which means that we won't unpack frozen output artifacts. This is mostly not noticable since it's pretty hard to get frozen output artifacts anyway

But it does also fix an issue where passing it would panic, instead of error. In the next diff I'm removing the incorrect code that allowed for the panic - I found this because this stopped compiling with that change

Reviewed By: IanChilds

Differential Revision: D82495397

fbshipit-source-id: c7f1c072c63adae43a4c3cce96f77d8d5167fc97