incremental: fix logic issue
Summary:
`writable_incremental_args` is created as a list of the objects. But then, instead of appending the  swiftdeps etc, it is recreated.

This is wrong - it's supposed to contain all the files that need to be made writable by swift_exec.py, so that we have an explicit list to pass in.

The only reason it works today is that swift_exec.py already has logic to make some files writable that is covering up this bug. (It probably should be removed).

This diff also preserves an `object_outputs` variable for use in other work (D82560637)

Reviewed By: NuriAmari

Differential Revision: D82648489

fbshipit-source-id: e89c72c5dee4e418dd9c64b2ade53f8783c7dd7b