dedupe concurrent get ttl calls
Summary:
Like it says in the title. This call is absurdly expensive in RE, and this leads to memory use that shoots up when we do a build where a lot of actions unblock at the same time and share a lot of directory paths.

This is expected to improve the exact same cases D82625781 improves, because if you have a lot of sharing in your directories, then you are going to also have a lot of sharing in your find missing calls, since:

- The memory use improved in that earlier diff lies in the directory structure.
- "Lots of memory won by sharing the directory structure" implies "lots of duplication in directory structure".
- We fetch the TTL of each of those blobs, so deduplicating calls will benefit if the directory structure of actions is highly duplicative.

Here is why I'm changing this:

This is Buck2's memory use after the earlier diff on the same build, when it hits 60GB:

https://fburl.com/scuba/stacks_to_scuba/cf2u41o9

Now it's a bit hard to see, so look at this screenshot. Circled in red is the list of digests held in memory by Buck2. The huge stacks that mention RE are the get digests ttl calls themsleves.

 {F1982052149}

I did not try to be very economical, or clever here. The existing call is very expensive, so there's a lot of margin for error here.

Reviewed By: JakobDegen

Differential Revision: D82639090

fbshipit-source-id: 77d54ff982b0e5b258f4f9653e438921e201bb18