make Uploader use RemoteExecutionClient, not REClient
Summary:
Like it says in the title. I want to be able to dupe the client, but I can to
do that if I have access to the higher level RemoteExecutionClient, which is an
`Arc` and can be duped (the internal `REClient` an actually be cloned, but not
the  OSS one).

So, refactor this code a bit so we pass in the higher level thing into
the uploader.

Reviewed By: ndmitchell

Differential Revision: D82719581

fbshipit-source-id: 1f846054d2828cf1ddd12c72e699ceb679523642