allow sharing per-command action paths via directory interner
Summary:
Like it says in the title. We don't currently share the directory trees
produced by actions, but when we run a lot of similar actions, there's quite
a bit of reuse to be had there.

Consider this build:

https://www.internalfb.com/sandcastle/job/13510801029693759

If you get a lot of cache hits this uses 20GB of RAM. If you get a lot of cache
misses this uses 100GB of RAM. That's too many sizes apart on Sandcastle, so as
a result that means that if you get a lot of cache misses, the job is doomed and will
fail 3 times in a row despite being repeatedly upsized.

This particular job is nice because it's really easy to repro. You just need to
add a comment in `fbcode/nodeapi/lib/core/privacy/zones/ZoneEvaluator.h` to
invalidate a bunch of builds, then you can build this set of targets to repro:
P1951200753.

Note 1:

- I do not know how prevalent this particular issue is. I do anticipate that
  with our new packing we try to set builds up for reuse, so not having reuse
  here is sad.
- I have made this configurable so we can actually tell what impact it has.

Note 2: the method is called "intern", but it's a bit of a misnomer: we don't
actually retain the directories shared this way forever. The "interner" holds
only weak references to the contents, so when the contents are released we just
end up with an index mapping to nothing. This diff does use a non-static
interner to ensure we also GC said index at the end of the build.

Note: the underlying directory and "interning" implementation is used for a few
things of importance in Buck2 such as literally every directory we retain.

Reviewed By: IanChilds, ndmitchell

Differential Revision: D82625781

fbshipit-source-id: db57dfc2634dd71478c0d1e0205beb52b7050027