make deduplication of concurrent get_digests_ttl calls configurable
Summary:
Like it says in the title. I want to be able to turn this on selectively so
that we can validate that it actually improves things. This also makes for
a safer rollout.

There is a bit of code-golfing here required to get the same behavior we were
getting before when it's off without just copy pasting all the code, since the
new implementation allocates a HashMap for digest-to-TTL lookups (it has to,
because now the response may contain more TTLs than we actually requested, so
we can't do the "just sort and zip" we did before), but the old implementation
could just sort-and-zip.

The way I made this work is:

- `process_get_digest_ttls_response` produces the sort-and-zipped iterator,
  and can operate on either references or owned digests.
- both branches produce an iterator of (digest-reference, ttl). The new one
  does it by taking the input digests and looking them up in a hashmap (which
  was produced using `process_get_digest_ttls_response` on owned digests), and
  the old one does it by just returning the output of
  `process_get_digest_ttls_response` directly (which is called on borrowed
  digests).

Outside of that there is just a million lines of threading a config through, as
one does.

Reviewed By: ndmitchell

Differential Revision: D82719582

fbshipit-source-id: 82db2422cbadfec4313bf14fb0f5a04a9c6ea22b