artifacts: Re-use allocated output artifacts in dynamic outputs attrs
Summary:
I apologize for this diff being a bit big, it wasn't super easy to split up.

The substantial change is here:

https://www.internalfb.com/code/fbsource/[e7ae03c3c802c51c74c8435dd5b1822067466ae7]/fbcode/buck2/app/buck2_action_impl/src/dynamic/params.rs?lines=40-41%2C48-49%2C52-55%2C64

Previously, we stored a list of `BoundBuildArtifacts` - we now instead store a list of `ValueTyped<'v, StarlarkOutputArtifact<'v>>`s. All the other changes are just shuffling conversions around to make that work

The reason to do this is memory: `BoundBuildArtifact` is 64 bytes, while the `ValueTyped` allows us to use only 8 bytes to hold a reference to the starlark value that we're allocating anyway

This also saves a bit more memory in another way, which is that now `DynamicAttrValue` is 40 bytes instead of 64

Reviewed By: IanChilds

Differential Revision: D82788948

fbshipit-source-id: 0cadb8fa759e6e8899cdb642d83ad2c2af0b7aa3