Use materializer to write action metadata
Summary:
We now (in buck2 in general) route almost all of our writes through the materializer, and it's much nicer to do so here too:
- We no longer have to have some special-case code to upload the action metadata to RE.
- That in turns means that we don't have to hold onto the `data` itself in `ActionMetadataBlob`.
- The materializer will take the `data` and compress it so we don't use as much memory while we're holding onto it.

This also makes it possible to use content-based paths for this input. Previously, when running locally we'd just overwrite the metadata path, now the materializer will handle all that correctly for us.

Reviewed By: scottcao

Differential Revision: D82811026

fbshipit-source-id: a8d61757c4ca8b9d696c290a5488cea58319858d