Filter out the nativeMain for boundle build
Summary:
When I upgrade the atomicu (the following diff), I noticed the nativeMain is not filter out when build the bundle releasee, which cause that diff reverted.
The error:
```
Exception in thread "main" java.lang.RuntimeException: File root/nativeMain/default/manifest has already been added!
```

After check code, I noticed we only applied it in here:
https://www.internalfb.com/code/fbsource/[19fa25a36c561e47361d6241e84d24a6c013ab72]/fbcode/buck2/prelude/toolchains/android/src/com/facebook/buck/android/apk/sdk/ApkBuilder.java?lines=699-703

but not in the bundle util part:
https://www.internalfb.com/code/fbsource/[19fa25a36c561e47361d6241e84d24a6c013ab72]/fbcode/buck2/prelude/toolchains/android/src/com/facebook/buck/android/bundle/AndroidBundleUtils.java?lines=369-372

It will be good we can use the same one, feel free to suggest how can we share the same function?

Reviewed By: hick209

Differential Revision: D82825398

fbshipit-source-id: a1af285b8d823ca5cd49bc1e83c96ed904098329