Add test for freezing and unfreezing
Summary:
This diff add the tests for freezing and unfreezing

The test script use mlock to allocate and lock memory in RAM to prevent the memory to be swapped from the memory.

Two actions will run in parallel to allocate and lock memory. One of the action will allocate until it detect the memory pressure, and it will be frozen. Other one will allocate and lock some memory and then sleep some time to simulate some on going work.

Reviewed By: Will-MingLun-Li

Differential Revision: D82766256

fbshipit-source-id: f875dfaf54dccf38c34670a94b2503608a273d9f