Add (inert) attribute for content-based cxx compile
Summary:
The attribute is threaded through the macro layer and the result stored in allow_content_based_paths, but this is NOT currently used. It will be enabled in the next diff.

Note: this is just the first part of the old D81327920, which has been split off, as per JakobDegen's suggestion.

Reviewed By: christycylee, milend

Differential Revision: D82957450

fbshipit-source-id: 9a6766d7968185b3981445ade54d6584fc0248b2