make the daemon aware of whether it has resource control enabled
Summary:
Like it says in the title. I want to expose the stats in a snapshot field, but
the stats are meaningless if the daemon does not have resource control, in
which case I do not want them to be published.

I do not want to rely on reverse engineering "do I expect this to be enabled?"
in the daemon since that's the kind of thing that always eventually breaks.

Reviewed By: Will-MingLun-Li

Differential Revision: D82817108

fbshipit-source-id: 130cee87bc22c734e314d787878cff4ac6d669fe