Add memory reading methods to CgroupInfo
Summary:
Expose a way to read memory.stat and do something useful with it.

Looking a memory.max or memory.current is ... not very insightful. This is inclusive of caches, so it sometimes means you are measuring the available memory, not the memory actually used. This is very true for the buck2 daemon, and at least partially true for local actons.

Reviewed By: Will-MingLun-Li

Differential Revision: D82557872

fbshipit-source-id: e62b2d16e55a30eb983c104afe5db8968b8f6659