Test for unbound artifact inside a TSet
Summary:
This is the same as the previous diff, except that if you "hide" the artifact inside a TSet, we don't actually do any verification that it is bound, which is not good.

Unfortunately it's also not clear to me how to fix this, am I missing something obvious? Anyone got any ideas?

Reviewed By: cjhopman

Differential Revision: D83050181

fbshipit-source-id: 13e825c79653dbfad965085ff309355af8f08788