timings: `TimeSpeed` to `Timekeeper`
Summary:
Ok, after a bit of flailing I finally figured out how I want to do this.

So, for context, I'd like to give `buck2 log replay` the ability to change speeds as it runs. It turns out that's pretty hard to do though at the moment, because the way we manage durations in superconsole is a bit of a mess (actually, that doesn't just affect replay, more on that in future diffs)

This stack so far is me figuring out how to clean that up - this diff sets up a dedicated component that will manage the superconsole's view of durations and virtual time, and can be used to ensure that these things all make sense

Reviewed By: IanChilds

Differential Revision: D83043101

fbshipit-source-id: e906e3c65f3de662cdcdf17667f4ac334c1eaa6a