Use content-based path hashing directly if turned on for Kotlin
Summary:
Previously, we couldn't use content-based paths as the output of an incremental action (wasn't correctly supported), so we would copy the outputs to a content-based path.

Now, we have support for content-based paths as the outputs of incremental actions, so for all output we just use the `kotlin_toolchain.allow_experimental_content_based_path_hashing` to specify whether to use content-based paths or not.

That's turned off right now, so there should be no effect on builds from this diff.

Reviewed By: cjhopman

Differential Revision: D83049539

fbshipit-source-id: 3534d6f1145fe3d6b1b984682048dfae485d5303