timings: Reimplement as a push-based replayer
Summary:
The replayer - the component that delays events to get `buck2 replay` timing to make sense - is currently implemented as a hand-written stream. This is cute, but is pretty gnarly to work with in practice and results in hard to read code.

Reimplement the thing as a push-based replayer. Specifically, we have a background task that just polls the event stream, does appropriate sleeps, and then sends the events onward

The way I've written this with the `loop` is slightly weird, but this is intentional to enable further changes

Reviewed By: cjhopman

Differential Revision: D83043098

fbshipit-source-id: f33c10e5dee5c9c59c495be3d0a15285ddc90c72