Rename uses_experimental_content_based_path_hashing -> has_content_based_path
Summary:
Not really a rename at this point, as I need to keep the previous param for compatibility.

I want to get rid of the `experimental` part as that's not really the case any longer, and I think `has_content_based_path` is a bit more succinct, but open to ideas.

Reviewed By: christycylee, Will-MingLun-Li

Differential Revision: D81583752

fbshipit-source-id: ef9035171e0681fbe73d03259f5a89b5b9731597