timings: Find zero time markers before starting superconsole
Summary: The way things are implemented right now, superconsole and the rest of the client are started before any parts of the event log are read. Instead, read the event log up until the first event before. This comes in useful in the next diff, and will be needed to implement `--seek` later

Reviewed By: cbnilrem, cjhopman

Differential Revision: D83043109

fbshipit-source-id: 7335a1d83496ebd923678b4b09948e97967eb59a