timings: Reorganize timekeeper around a clock
Summary:
If we are going to allow changing speeds during replay, the replayer and superconsole need to be able to talk to each other somehow. For that reason, and other reasons, introduce a clock component into the timekeeper that can take some responsibility for this.

For now the clock is extremely simple and implemented the same everywhere, next step is to do more interesting things with it.

One good use of this would be to write a `TestClock`. I also don't know what's going on with `what-up` - it can't be right - but this can probably be used to fix that too

Reviewed By: cjhopman

Differential Revision: D83043110

fbshipit-source-id: b202dbe9dbecd5711b6f4513b938835848f7a6fa