Unset and restore memory.high in frezzing logic
Summary:
This diff adds support Unset and restore memory.high in frezzing logic.

When frozen_cgroups is not empty and is under pressure, we unset the memory.high to max

when frozen_cgroups is empty and is not under pressure, we restore the memory.high back to the original setting.

Reviewed By: Will-MingLun-Li

Differential Revision: D83071248

fbshipit-source-id: 964ea129f2486124451714cba19116fdd08cb81a