Refactor the ResourceControlRunnerConfig
Summary:
This diff refactor the type `ResourceControlRunnerConfig` to have three variant:
1. BuckDaemon
2. ActionRunner
3. Forkserver

It make the configs more clear.

Also **what's more important** here is that old implementation have an issue that it also set the memory.high and memory.max for the daemon, which is not what it meant to do as the description of the config.

We already set the memory.high and memory.max at here for the cgroup for daemon, forkserver and workers

https://www.internalfb.com/code/fbsource/[038b1c54ec37]/fbcode/buck2/app/buck2_client_ctx/src/daemon/client/connect.rs?lines=522

Reviewed By: Will-MingLun-Li

Differential Revision: D83003529

fbshipit-source-id: b081a3f4aab64bc94313a3df58c3bb36fefddb23