timings: De-duplicate serde impls
Summary:
We have two `Invocation` types - the `buck2_data` one and one in `buck2_event_log` with slightly more ergonomic types. Right now, both of these impls serde traits. We're lucky for now that the serde impls from these are compatible with each other, but I almost wrote a bug in the next diff that resulted in these not being compatible

Instead of requiring them to be compatible, this diff removes the serde impls in `buck2_event_log` and just uses the ones on the proto types

Reviewed By: cbnilrem

Differential Revision: D83043122

fbshipit-source-id: 5146be9257b5cf4f01d56182a0e33e97955c4bf3