timings: Use event timings instead of client timings for `time_` fields
Summary:
The `time_to_` columns currently all use an end time computed in the invocation recorder in the client. That means that in scenarios with potential event lag, these timings will disagree with what's happening in the daemon. At the same time, other telemetry that we emit (such as all of the `duration_ms` columns) is based on the event stream and so contains daemon timings. That all seems like a footgun, just use the daemon timings here too

It feels a bit sad that grepping for `::now()` in the client leaves you with a pretty good chance of finding bugs

Reviewed By: cjhopman

Differential Revision: D83043127

fbshipit-source-id: 8b74f5e014f5f34a3e88a74f222073a7b6024ca8