timings: Fix delay computation in event debugger
Summary:
D45605221 introduced a bug. It changed the `start_time` that was being passed from being the start time of the command to the start time of the current tick. Besides the code that I deleted in the previous diff, the primary thing that this was being used for was to calculate in the event debugger whether there was a delay in the events making it to the client, and it did actually want the start time for that:

https://www.internalfb.com/code/fbsource/[54270c72232c4bc1c6bff08a36594c2ccd01343b]/fbcode/buck2/app/buck2_event_observer/src/debug_events.rs?lines=69-84

You could generally tell that something suspicious was going on here from the part where this variable was named `start_time` in some places and `received_time` in others

Afaict the event timing delay computation has been broken since then as a result.

This diff fixes it and simplifies the whole thing by simply computing the delay since when the event was generated in the event handler. Making this thing impure isn't *ideal*, but it's certainly an improvement over what we had before

Reviewed By: cjhopman

Differential Revision: D83043106

fbshipit-source-id: f99c292858aa7917586d0efda12a318c9e875f7a