timings: Use actual event timings in `EventTimestamp`
Summary:
First of all, I promise I really I am trying my best with reviewability here, unfortunately I am finding it hard to edit this code without keeping all of it in my head at the same time, so possibly reviewers will be subjected to similar constraints

The main thing that this diff does is change `EventTimestamp` from holding an `Instant` representing the client's computation of when the event was received, to instead hold the actual realtime at which the event was sent by the daemon. This has a myrhiad of consequences, in no particular order:

 1. The tests no longer depend on their execution timing for getting the right output. The event timestamp and timekeeper are both set to controlled values. This means we delete the time dilation crap there
 2. The span durations reported in superconsole are now based on the system realtime clock (`SystemTime` in Rust, the thing that's generally encoded by `prost_types::Timestamp`), not the system monotonic clock (`Instant` in Rust). This has some unimportant consequences in cases where you like change the system time in the middle of a build. A slightly more important consequence is on systems (Macs) where the monotonic clock pauses when the system sleeps. In these cases, when woken back up superconsole would previously keep counting as if the sleep had never happened. Now it'll reflect all the slept time. This seems fine.
 3. The important one - if there is a delay between the event being produced by the daemon and being received by the client, the time shown in superconsole will reflect the time since the daemon produced the event. In this case, spans may show up in superconsole with non-zero starting times. That seems likely to be an improvement though, since it means that the times better reflect how long something has actually taken and also it'll allow us to know when this is happening.
 4. The math that the timekeeper does is now generally in the timeline of the event stream. That means we no longer need to pass the speed to the timekeeper explicitly, and all the speed related code is now restricted to the replay implementation (this is the thing that unblocks implementing speed changing)

The code that computes an instant when receiving an event is not gone yet - it's still used for one other thing. Edit: I successfully removed this two diffs later

I'll probably post about this when landing it

Reviewed By: cjhopman

Differential Revision: D83043103

fbshipit-source-id: c5cd85e9fc08e7590fd1186b9bf6927dcabd78c5