do not report `null` for max memory when no actions were observed
Summary:
There's an inconsistency in the reporting of detailed build metrics: if we
ran no local actions, then the runtime of local actions is zero, but their
memory use is ... null.

Downstream in CI that's a little annoying because we want to differentiate "we
do not know what the max memory use is" and "we know there are no local
actions", but right now both are reported as null.

I actually had a diff to fix this in CI, but it seems logical to actually fix
it in Buck2 by making the reporting consistent with other metrics, so here
we are.

Reviewed By: IanChilds

Differential Revision: D83235826

fbshipit-source-id: 0c20431a846445c86806bc6f248ed876a6331e17