timings: Report non-zero span timings in whatup
Summary:
Whatup has the behavior of always reporting zero for durations, as shown in the tests in the previous diff. This is now obvious in the code as the result of it just blindly using the realtime clock with poorly chosen timestamps.

This diff broadly fixes the time handling in whatup, using a suitable clock and slightly better start timings, thereby letting durations take on the value the user would expect

Reviewed By: IanChilds

Differential Revision: D83043124

fbshipit-source-id: b9cdbf349f540bc797848c7d09aebe28dd7249fd