timings: Use start_time as base of super console command duration
Summary:
As mentioned in the last diff, superconsole displays a "command duration" - the time since the command started. However the "command started" in the sense there is currently just "the ticker is initialized" which is a pretty arbitrary point in the client's lifetime, and also is not logged meaning it can't be reproduced in replays.

Instead, use the actual command start time, which is now logged and can be reproduced, and also is a closer approximation of the user's experience.

This does mean that if there's a delay between the user's build starting and superconsole starting, the command duration displayed there will "start" at a non-zero value. That seems good though

Reviewed By: IanChilds

Differential Revision: D83043132

fbshipit-source-id: 2f5e07004d32431c5f4a98ae2eb76fa64ec562e8