Include third-party framework deps in swiftinterface compilation
Summary:
## The problem

Currently, using explicit modules with third-party frameworks that contain Swift is broken.
Consider this:
```
apple_library(name = "Foo", deps = [":LibraryA:"])
prebuilt_apple_xcframework(name = "LibraryA", deps = [":LibraryB"])
prebuilt_apple_xcframework(name = "LibraryB")
```

Before compiling `Foo`, Buck attempts to compile the Swiftinterface for `LibraryA`. But despite an explicit dependency on `LibraryB`, it does not compile the Swiftinterface for `LibraryB` and include it in the action Therefore if `LibraryA`'s Swiftinterface contains `import LibraryB` it will fail.

(An exception to this is if `LibraryB` is mixed Swift + Objective-C, in which case the errors can become more obscure because they will be limited to the Swift portion, but the import will likely succeed).

## This diff

Buck attempts to build `LibraryA`'s binary swiftmodule because the providers are already there. However, `compile_swiftinterface_common` only includes the SDK dep providers (which are in `WrappedSdkCompiledModuleInfo`). This diff also includes the `SwiftDependencyInfo` that includes the swiftinterfaces from the third-party deps.

Reviewed By: fabiomassimo

Differential Revision: D83238510

fbshipit-source-id: e63defd5307760905f0c89ba43e21ffb01e1caff