replay: Fix support for old log formats
Summary:
In D83190556 I broke a bunch of tests. The reason is that I started using the invocation record's timestamp as the "zero time" of a replay, but in logs with that field missing, it was defaulted to the unix epoch, and so the replay just sat there waiting for 1970 -> now to pass. Fix this by correctly accounting for the possibility of the timestamp being missing

With what-up this is much less important, since there it only affects timestamps in the output and not whether the thing runs at all, so skipped it there

Reviewed By: scottcao

Differential Revision: D83442690

fbshipit-source-id: c29475af1a2aaa86052b8227c6ff0eaedd72c9e9