Skip hidden artifacts when creating artifact_path_mapping for RunAction
Summary:
We create the `artifact_path_mapping` so that we can resolve any paths when doing `add_to_command_line`. If we have hidden artifacts, those are never added to the command line, so we don't need a mapping for them. So, we collect just the artifacts that are non-hidden, and then create a mapping for those.

Meta: in particular, header maps are huge but are hidden, so that's where a bunch of the wins come from here.

Reviewed By: scottcao

Differential Revision: D83339746

fbshipit-source-id: 73558c6587b4de19d0d2e879a18bcd02ccafcdc4