Get critical files for source dbs
Summary:
This diff updates `pyrefly.bxl` to query for build files that affect type checking, so that we can add it to our file watchers.

Alternatively, we could just have our `BuildSystem` just mark a few file names as critical file types, and run an `invalidate_config()` every time we see an edit on one of them.

Reviewed By: grievejia

Differential Revision: D82758043

fbshipit-source-id: 39b41410057491be4ad70b5b72f3dcff0e594046