Support for explicit Git object formats in .buckconfig
Summary:
This commit adds a new optional config option for external cells
retrieved via Git:

    [external_cell_foo]
    git_origin = ...
    commit_hash = ...
    object_format = sha1

    [external_cell_bar]
    git_origin = ...
    commit_hash = ...
    object_format = sha256

If specified, we pass `--object-format sha1` or `--object-format sha256`
as arguments to the `git init` invocation.
In both cases we validate that the commit_hash is a valid digest for the
respective algorithm. If no object_format is given, we check that the
commit_hash is a valid SHA1 digest, but we do not pass the
--object-format option when initializing the Git repository.