Resource broker now prefers the default device
Summary:
Currently when setting up a simulator to run Apple tests, if there's no specific device requested by a config, the resource broker will select **any** existing simulator for the newest available runtime regardless of the device.

This creates problems where a test requires a specific device (usually screenshot tests) and the user has an existing simulator for another device.

This diff changes the resource broker to **require** the a simulator for the default device when a device isn't specifically set with a config.

This means that test runs will always match the device used in CI unless the user specifically overrides it.

Reviewed By: mark-cmd-alt

Differential Revision: D83497761

fbshipit-source-id: faf83dd22e08ae4b963c0a7ae95ce50dd4c8faf3