Enable transformation spec for rust with static linking
Summary:
Transformation specs should also be available for Rust (mostly to migrate off link group-based APIs that is currently rolled out for rust too). Thankfully this is easy to configure because all we should do is to propagate the values and all handling will work out of the box.

Passing transformation spec context to part that handles static linkage

Reviewed By: dtolnay

Differential Revision: D83559463

fbshipit-source-id: bf7f71152cec516c3d12d2a36657f0cb925b0520