Save content-based incremental state after cache hits too
Summary: This fixes the bug described in the last diff. I'm not a huge fan of how we have to update the incremental state manually for every different possible way of executing an action (rather than doing this just once in `run::execute_inner`), but I know Will-MingLun-Li had some reasons why it was better for the incremental state to be in `buck2_execute_impl` so this might just have to do I guess.

Reviewed By: Will-MingLun-Li

Differential Revision: D83650014

fbshipit-source-id: 27f1feb929ee1aed2ec0ac21cc4edf68e818b360