Write test to show bug in incremental handling on a cache hit
Summary:
There's a bug (fixed in next diff) where if we get a cache hit, then we don't update the incremental state for content-based outputs.

This means that when the user rebases and gets a fully-cached build, the next time the actions is executed, we don't do it incrementally, which is a regression.

Reviewed By: Will-MingLun-Li

Differential Revision: D83649500

fbshipit-source-id: 93037b239eea315eec2fee51d1f43fbab9a9aeab