do not strip property initializers
Summary:
Previously we used to strip property initialisers, because they were causing problems for the target prototype was built on.
For example initalizer was causing problems for:
https://www.internalfb.com/code/fbsource/[f82ed475f55cac4262f5b13b558f14b7e77b6a5e]/fbandroid/java/com/instagram/creation/capture/assetpicker/cutout/util/CutoutBitmapUtil.kt?lines=310

val DEFAULT_SHADOW_COLOR: Int = Color.argb(128, 0, 0, 0)

But when we set SKIP_BODIES flag correctly, these initalizers are not getting handle by IR and don't need to be stripped.

However, if we strip initializers, we loose default values for constants or set incorrect default values

Reviewed By: siaojiecai

Differential Revision: D83738360

fbshipit-source-id: 894a0147d2cdb638d520dacd9b6df066088c452d