Add support for common buck2 params on test-shell via an application environment
Summary:
# Context
The buck2 test-shell uses buck2 to rebuild the code being tested and eventually reload it.

# Problem
At the moment there is no way to give it specific modifiers, so that if any constraints are needed to rebuild correctly (e.g. for debug mode), they will get lost

# This diff
As a first step towards fixing this, we add the possibility of specifying extra buck2 arguments via an application environment variable.

We reuse the (false) app name currently used for specifying the "search module", and factor out the logic

Reviewed By: michalmuskala

Differential Revision: D83584210

fbshipit-source-id: 4fe8c635207fa233bcc8d5760047827697399485