Add ability to unfreeze earlier
Summary: Currently, we only unfreeze once there are no actions running which is fine for experimental purposes, but we are definitely wasting resources and regressing performance by more than necessary. Add the ability to unfreeze earlier if the total memory used by buck2.slice is less than the memory when an action was frozen at + total memory last time something was frozen. This algorithm can still be further improved but it's a start

Reviewed By: cbnilrem

Differential Revision: D83712529

fbshipit-source-id: a46d7e345901b956a0bd3b2dfc6a4982a6fd1e4c