directory: Set up a/b test for directory merging changes
Summary:
The rest of this stack makes a number of changes to how directory merging works. These have significant perf impact, so let's enable ourselves to a/b test them. This diff forks the main piece of code that is going to be affected by this

The config is defaulted to true everywhere and then disabled in fbsource - that ensures that all the new code gets exercised in our tests

Reviewed By: IanChilds

Differential Revision: D83725662

fbshipit-source-id: fd31cf5586dba111a57816d764fd7c450c57445c