directory: Reproduce inefficient directory reuse when merging
Summary:
Consider a python library like

```
python_library(
    deps = [":bigdep", ":util"],
)
```

Where `:util` is already a dep of `:bigdep`, and what happens when you create the action directory for the source artifacts. If the python dep graph and directory structure works out so that the source artifacts for `:util` are `fbcode/util/p.py` and for `:bigdep` `[fbcode/bigdep/p.py, fbcode/other/p.py, fbcode/util/p.py]`, this all kind of works out ok. But imagine instead that `:util` has `[fbcode/a/simple.py, fbcode/b/simple.py]` while `:bigdep` has `the same + [fbcode/a/complex.py, fbcode/b/complex.py]`

What we should be doing in this case is recognizing that `fbcode/a` in `:util` is a subset of `fbcode/a` in `:bigdep`, and reusing the `ImmutableDirectory` from `:bigdep`. We don't however, and this test adds a test that shows that.

Effectively, this means that we'll rebuild and rehash the entire merkle tree for `:util` when doing this merge, up until the point where `:util` and `:bigdep` are actually completely identical. That's quite expensive.

Reviewed By: IanChilds

Differential Revision: D79666800

fbshipit-source-id: 16a014ad5dbf9d8c0f6153ffa2c33c26a64787f5