directory: Refactor directory merging to delay `.as_mut()` calls
Summary:
Prepare to resolve the issue from the previous diff - we need to avoid calling `.as_mut()` if the merge is not actually going to introduce anything new. However, we cannot know that until after having recursively merged children, so we need to refactor this logic to call `.as_mut()` only at that point

It wasn't easy to get this to be super clean. I did my best to balance readability and conciseness, not generalizing too much or too little, but happy to take feedback

There's one thing here - `Operation::Keep` - which is unused so far. I'll use it later in this stack, it's easiest to introduce it here to keep most of the complicated logic in one place

Reviewed By: cbnilrem

Differential Revision: D79666801

fbshipit-source-id: 566358e809f54249136c0c9831df3af38b4c4d90