Remove JavaVersion enum
Summary: Removing enum as it can easily be represented as an int. This file needs constant updating when Java versions are added and creates overhead to upgrading to new versions. We can just parse the version to an integer everywhere this is used to aid in making the java tool chain more fluid as we add new versions.

Differential Revision: D83854420

fbshipit-source-id: 3b93de051d271e45e67ded63f5bf0277d3716a24