directory: Add a `LazyDirectoryBuilder`
Summary:
So, there was an optimization that I initially applied to just tset merging, but in looking at profiles it's actually applicable in a bunch of other places as well. To be able to apply it more broadly, create a helper type that will apply the optimizations in question

For now this does nothing, I'm going to use this first before optimizing it, that way the perf win is in a single diff and easier to benchmark

I did have to add a trait to abstract over the builders. I'm always a bit sorry when introducing something like that to code that already has too many generics floating around, but fortunately I succeeded in keeping the trait goop to a bare minimum in the next diff

Reviewed By: IanChilds

Differential Revision: D83725659

fbshipit-source-id: d8ef04d48f282c7e3287abfc9e387ffa9e3b2912